Fix reported bugs

+++ app/src/main/java/ir/kindnesswall/view/main/conversation/chat/ChatAdapter.kt
@@ -102,7 +102,7 @@ class ChatAdapter(val viewModel: ChatViewModel) : RecyclerView.Adapter<RecyclerV
             is OthersChatsViewHolder -> {
                 holder.binding.item = item as TextMessageModel
                 if (!item.ack) {
-                    viewModel.sendAckMessage(item.id, item.chatId.toLong())
+                    viewModel.sendAckMessage(item.id, item.chatId)
                     item.ack = true
                 }
             }
